Setup CMake project + small test for libs
I am purposefully leaving all external source files in rather than excluding them in the gitignore in case I want to make any changes or fixes to the libraries.

Setup a cmake project with a temporary folder structure and some files I'm expecting to use (empty for now)

I've also created a small test in main to make sure the libraries work and compile, this test project creates a window and plays a sine wave on closure of said window.