Added some visualization and the ability to play audio from files and trough microphone
I've added a waveform and a spectrogram to visualize the audio.
To test both I also added the ability to play audio trough a wav file and using your microphone, the won't really be used later on and are purely for testing purposes and because it's just kinda cool to play around with.

The parser for the wav files is very simple and only supports 16-bit PCM WAV files, but no more is needed for our testing purposes.